+++ lib/src/controls.dart
@@ -177,9 +177,15 @@ class _BottomBarState extends State<BottomBar> {
                         : Icons.fullscreen,
                     color: Colors.white,
                   ),
-                  onPressed: () => controller.value.isFullScreen
-                      ? controller.exitFullScreen()
-                      : controller.enterFullScreen(),
+                  onPressed: () {
+                    if (controller.value.isFullScreen) {
+                      controller.exitFullScreen();
+                      triggeredFullScreenByButton = false;
+                    } else {
+                      triggeredFullScreenByButton = true;
+                      controller.enterFullScreen();
+                    }
+                  },
                 ),
         ],
       ),