+++ CHANGELOG.md
@@ -1,7 +1,9 @@
-## 3.0.0-beta.1
+## 3.0.0-beta.2
+* **BREAKING CHANGES** See Migrating to 3.x.x for detail.
 * Faster fullscreen toggling.
-* Toggles fullscreen on orientation change.
+* Toggles fullscreen on orientation change. [Issue #3](https://github.com/sarbagyastha/youtube_player_flutter/issues/3).
 * Option to add own custom thumbnail.
+* **FIXED** Video stuck at unknown state [Issue #24](https://github.com/sarbagyastha/youtube_player_flutter/issues/24).
 
 ## 2.0.0
 * **FIXED** iOS playback is fully functional. [Issue #2](https://github.com/sarbagyastha/youtube_player_flutter/issues/2)